## srmech 0.9.0rc7 — the Q carrier

The numpy-removal arc gave the package its numpy-free **array** carriers
(`Mat`/`Vec`/`HV`); the scalar was still rotating to `float`
mid-cascade. This rc closes that leak with **`Q`** — a reduced `(num,
den)` integer pair (`srmech/amsc/q.py`) — so a transcendental stays in
the integer ALU all the way and `float(q)` is the *one* last rotation,
taken only at the display / carrier edge (ALU-all-the-way /
FPU-last-mile; F868 stay-rational).

### What changed
- **Nine transcendentals flip `float → Q`**
(`srmech.amsc.rational.{sin,cos,tan,atan,atan2,exp,log,sqrt,hypot}`).
`cos²θ+sin²θ` is *exactly* `1`; `sqrt(4)==Q(2,1)`; `hypot(3,4)==Q(5,1)`.
`cexp`/`complex_exp` still return the builtin `complex`. The new
`rational.log` (was missing) is wired into the
`calculus`/`asymptotic_calculus` aliases. Non-finite input now
**raises** (a `Q` can't be ±inf/nan; overflow/underflow gates dropped);
`log(x≤0)` raises.
- **Native Q61 C peers** `srmech_{sin,cos,atan}_q61` +
`srmech_{exp,log,sqrt}_q61` (thin wrappers over the proven cores) expose
the int64 Q61 *before* the float projection; bound in `_native.py` and
dispatched from `rational.py` when native is present. **Byte-exact**
with the pure cascade (`test_native_q61_parity_rc7.py`). ABI stays **3**
(additive symbols).
- **Consumer normalization (~40 sites)** — flow `Q` in exact contexts,
collapse to `float` at genuine FPU boundaries: physical-observable leaf
ops (`qm/`, `kepler`) and iterative numeric kernels
(Jacobi/QR/SVD/Fiedler, Kuramoto, sigproc taper/window helpers), whose
rotations are irrational (a 3×3 Jacobi *hung* in `Q`; 1.4 ms after
collapse). `_fsqrt`/`_fhypot` made **libm-faithful on the non-finite
domain** the sweeps reach (`sqrt(+inf)=+inf`) rather than inheriting the
leaf's stay-rational raise.
- **Carrier docs** — README "Carriers" section + research-notebook §3.39
document the five-carrier family (`Mat`/`Vec`/`HV`/`Q`/`Complex128`),
the ALU/FPU lens, and the per-function A–N cascade reading.

### Invariants
`tools.total` stays **317** (the 9 ToolEntry return types flip `float →
Q`; no op add/remove). ABI **3**. numpy-free. MIT. 5-SSOT bumped
`0.9.0rc6 → 0.9.0rc7`.

+++ docs/srmech/python/srmech/amsc/cascade/hypercomplex_dft.py
@@ -84,7 +84,7 @@ _MU_AXES = {
 }
 # The body-diagonal unit axis (i+j+k)/√3 — the order-3 (triality-adjacent)
 # pure-quaternion direction; still μ²=−1.
-_S3 = 1.0 / _rsqrt(3.0)
+_S3 = 1.0 / float(_rsqrt(3.0))
 _MU_AXES["ijk"] = (0.0, _S3, _S3, _S3)
 
 _FORMS = ("left", "right")
@@ -140,7 +140,7 @@ def _resolve_mu(mu_axis, *, octonion) -> List[float]:
             return _as8(_MU_AXES[mu_axis])
         if mu_axis == "diagonal":
             hi = 8 if octonion else 4
-            inv = 1.0 / _rsqrt(float(hi - 1))
+            inv = 1.0 / float(_rsqrt(float(hi - 1)))
             return [0.0] + [inv if 1 <= i < hi else 0.0 for i in range(1, 8)]
         raise ValueError(
             f"mu_axis must be one of {sorted(_MU_AXES) + ['diagonal']}, or a unit "
@@ -155,7 +155,7 @@ def _resolve_mu(mu_axis, *, octonion) -> List[float]:
             "a quaternion mu_axis must lie in ℍ (components e4..e7 == 0); use "
             "octonion_dft / a quaternion-scope coupler for an octonion axis"
         )
-    norm = _rsqrt(float(sum(c * c for c in v[1:])))
+    norm = float(_rsqrt(float(sum(c * c for c in v[1:]))))
     if norm == 0.0:
         raise ValueError("mu_axis must be a non-zero pure-imaginary vector")
     inv = 1.0 / norm
